+++ db/migrate/20160328220517_add_credits_to_plans.rb
@@ -1,5 +1,5 @@
 class AddCreditsToPlans < ActiveRecord::Migration
   def change
-    add_column :plans, :credits, :integer
+    add_column :plans, :credits, :decimal
   end
 end
